% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustmentFunctions.R
\name{adj_plot}
\alias{adj_plot}
\title{Adjustment Radar Chart -- adj_plot()}
\usage{
adj_plot(wimp)
}
\arguments{
\item{wimp}{Subject's WimpGrid. It must be a "wimp" S3 object
imported by  the \code{\link{importwimp}} function.}
}
\value{
A Plotly radar polar plot.
}
\description{
This function creates a radar chart showing the value of the
             self for each construct and its adjustment with respect to the ideal.
}
\examples{

adj_plot(example.wimp)

}
\author{
Alejandro Sanfeliciano
}
