% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentralityFunctions.R
\name{betw_index}
\alias{betw_index}
\title{Betweenness Centrality Index -- betw_index()}
\usage{
betw_index(wimp, norm = TRUE)
}
\arguments{
\item{wimp}{Subject's WimpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}

\item{norm}{If TRUE, values will be normalized. Default is TRUE.}
}
\value{
Returns a vector with the betweeness index for each of the
        constructs.
}
\description{
Calculates betweenness centrality (number of shortest paths
             passing through each construct).
}
\examples{

betw_index(example_wimp)
betw_index(example_wimp, norm = FALSE)

}
\author{
Alejandro Sanfeliciano
}
