% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentralityFunctions.R
\name{close_index}
\alias{close_index}
\title{Closeness Centrality Index -- close_index()}
\usage{
close_index(wimp, norm = TRUE)
}
\arguments{
\item{wimp}{Subject's WimpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}

\item{norm}{If TRUE, values will be normalized. Default is TRUE.}
}
\value{
Returns a vector with the closeness index for each of the
        constructs.
}
\description{
Calculates closeness centrality of constructs within the
             implication digraph (inverse of average shortest distance).
}
\examples{

close_index(example_wimp)
close_index(example_wimp, norm = FALSE)

}
\author{
Alejandro Sanfeliciano
}
