% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentralityFunctions.R
\name{construct_colors}
\alias{construct_colors}
\title{Construc color matrix based on its category}
\usage{
construct_colors(wimp, mode)
}
\arguments{
\item{wimp}{An object of class 'wimp', which contains an implication grid
and associated constructs.}

\item{mode}{Color mode for .color.selection (i.e, "red/green")}
}
\value{
A matrix containing construct colors based on its category and .color.selection.
}
\description{
This function computes the presence (P, frequency of occurrence) and
hierarchy (H, influence on others) indices for constructs within an implication grid.
It can standardize these indices based on the maximum degree if required.
}
