% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustmentFunctions.R
\name{construct_index}
\alias{construct_index}
\title{Construct Congruence Analysis -- construct_index()}
\usage{
construct_index(wimp)
}
\arguments{
\item{wimp}{Subject's WimpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}
}
\value{
A matrix with frequency and proportion values for each construct
        type: Congruents, Discrepants, Dilemmatics, and Undefined.
}
\description{
Calculates frequency and proportion of construct types
             (congruent, discrepant, dilemmatic and undefined) based on
             self-ideal relationship analysis.
}
\examples{
construct_index(example_wimp)

}
\author{
Alejandro Sanfeliciano
}
