% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentralityFunctions.R
\name{constructs_dendrogram}
\alias{constructs_dendrogram}
\title{Constructs Dendrogram}
\usage{
constructs_dendrogram(wimp)
}
\arguments{
\item{wimp}{A `wimp` object containing the constructs and the corresponding
scores in the P-H vector space. This object should have been generated using
the GridFCM.practicum package.}
}
\value{
A dendrogram visualizing the hierarchical clustering of constructs
based on their distances in the P-H space. The dendrogram highlights the
optimal clustering of constructs.
}
\description{
This function generates a dendrogram of constructs based on the Euclidean
distances in the P-H vector space. It automatically calculates the optimal
number of clusters for grouping the constructs.
}
