% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtCFunctions.R
\name{cyr_index}
\alias{cyr_index}
\title{Calculation consequence and feedback}
\usage{
cyr_index(wimp, std = FALSE)
}
\arguments{
\item{std}{}
}
\description{
Calculation consequence and feedback
}
\examples{
cyr_index (example.wimp)

}
