% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentralityFunctions.R
\name{degree_index}
\alias{degree_index}
\title{Degree Centrality Index -- degree_index()}
\usage{
degree_index(wimp, method = "weight")
}
\arguments{
\item{wimp}{Subject's WimpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}

\item{method}{Centrality calculation method: "simple", "norm", "weight",
"wnorm", or "ego". Default is "weight".}
}
\value{
Returns a list with the centrality data by construct and separated by
        input degree, output degree and total degree (in and out).
}
\description{
Calculates centrality based on construct connections. Centrality
             represents the degree of connection each construct maintains
             with others (number of links per vertex).
}
\examples{

degree_index(example_wimp)

}
\author{
Alejandro Sanfeliciano
}
