% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentralityFunctions.R
\name{dismatrix}
\alias{dismatrix}
\title{Shortest Distance Matrix -- dismatrix()}
\usage{
dismatrix(wimp, mode = "out")
}
\arguments{
\item{wimp}{Subject's WimpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}

\item{mode}{Distance calculation mode: "out" (respecting edge direction),
"in" (inverse direction), "all" (ignoring direction). Default is
"out".}
}
\value{
Returns the digraph distance matrix. Matrix that contains the
        distances of the shortest paths from one construct to another.
}
\description{
Calculates shortest distances between construct pairs in the
             implication digraph.
}
\author{
Alejandro Sanfeliciano
}
