% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentralityFunctions.R
\name{eigen_index}
\alias{eigen_index}
\title{Eigenvalue Centrality Index -- eigen_index()}
\usage{
eigen_index(wimp, matrix = "weights", num_vectors = 2)
}
\arguments{
\item{wimp}{An object of class 'wimp' (weighted implications grid).}

\item{matrix}{Matrix type for analysis: 'direct', 'weights', or
'implications'. Default is 'weights'. Note: only 'weights' is
available in new format.}

\item{num.vectors}{Number of eigenvectors to use for centrality computation.}
}
\value{
A dataframe containing the constructs' names and their respective
        centrality scores.
}
\description{
Calculates centrality scores based on eigenvalue decomposition
             of the adjacency matrix.
}
\examples{

eigen_index(example_wimp)

}
\author{
Carlos Hurtado
}
