% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R
\docType{data}
\name{example.wimp}
\alias{example.wimp}
\title{Example WimpGrid Object}
\format{
An object of class \code{wimp} (inherits from \code{list}) of length 6.
}
\usage{
example.wimp
}
\description{
A sample "wimp" object used as an example dataset in the package.
             This object contains constructs, self, ideal matrices, and other
             relevant data for testing and demonstration purposes.
}
\examples{
digraph(example.wimp)
}
\keyword{datasets}
