% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R
\docType{data}
\name{example_wimp}
\alias{example_wimp}
\title{Example WimpGrid Object}
\format{
An object of class \code{wimp} (inherits from \code{list}) of length 3.
}
\usage{
example_wimp
}
\description{
A sample "wimp" object used as an example dataset in the
             package.This object contains constructs, self, ideal matrices,
             and other relevant data for testing and demonstration purposes.
}
\examples{
digraph(example_wimp)
}
\keyword{datasets}
