% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentralityFunctions.R
\name{graph_ph}
\alias{graph_ph}
\title{Scatter Plot of Wimps Constructs in P-H Space}
\usage{
graph_ph(..., mark.nva = TRUE, mark.cnt = TRUE, show.points = TRUE)
}
\arguments{
\item{mark.nva}{Boolean value that specifies if non-viable areas are to be marked in the
graphic. Non-viable areas are parts of the P-H space where constructs cannot logically exist.}

\item{mark.cnt}{Boolean value that specifies if central constructs have to be highlighted.
Central constructs are depicted with a distinct color and symbol to differentiate them from
peripheral constructs.}

\item{show.points}{Boolean value that specifies whether points should be displayed or not}

\item{wimp}{A `wimp` object containing the data from which the PH matrix
is derived.}

\item{method}{A character string specifying the method used for calculating
the PH matrix. Default is `"weight"`.}

\item{std}{A logical value indicating whether the data should be
standardized before calculating the Mahalanobis distance. Default is `FALSE`.}
}
\value{
A Plotly object representing the generated scatter plot.
}
\description{
This function generates a scatter plot of constructs of a given wimp in the P-H space,
where P represents Presence (frequency of the construct) and H represents Hierarchy
(influence of the construct). Central constructs are highlighted in red color, and
peripheral constructs in another, facilitating their visual identification.
}
