% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{idealdigraph}
\alias{idealdigraph}
\title{Ideal digraph -- idealdigraph()}
\usage{
idealdigraph(wimp, ...)
}
\arguments{
\item{wimp}{Subject's WimpGrid}

\item{...}{Additional arguments}
}
\description{
Plot the ideal self based on the constructs and their relations.
}
