% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{inout_digraph}
\alias{inout_digraph}
\title{In - Out digraph -- inout_digraph()}
\usage{
inout_digraph(wimp, iso.con, ...)
}
\arguments{
\item{wimp}{Subject's WimpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}

\item{iso.con}{Objective construct to analyse the in and out. Must be an
integer representing its position in the list of constructs.}

\item{...}{additional arguments are passed from \code{\link{digraph}}
function.}
}
\value{
A digraph made with visNetwork
}
\description{
A digraph showing the in and out vertices for a given construct.
}
\examples{

idealdigraph(example.wimp)
}
\author{
Alejandro Sanfeliciano
}
