% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentralityFunctions.R
\name{mahalanobis_index}
\alias{mahalanobis_index}
\title{Calculate Mahalanobis Indices for Constructs}
\usage{
mahalanobis_index(wimp, method = "wnorm", std = "none", sign.level = 0.2)
}
\arguments{
\item{wimp}{A `wimp` object containing the data from which the PH matrix
is derived.}

\item{method}{A character string specifying the method used for calculating
the PH matrix. Default is `"wnorm"`.}

\item{std}{A logical value indicating whether the data should be
standardized before calculating the Mahalanobis distance. Default is `FALSE`.}
}
\value{
A matrix that includes the original PH matrix with two additional
  columns: one for the Mahalanobis distance of each construct and another
  indicating whether the construct is considered "central" based on the
  chi-square cutoff.
}
\description{
This function calculates the Mahalanobis distance for each construct
in a given PH matrix obtained from a `wimp` object. It also determines
whether each construct is considered "central" based on a chi-square
cutoff, which is calculated using a specified significance level.
}
\examples{
result <- mahalanobis_index(wimp, method = "weight", std = FALSE)
head(result)

}
