% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentralityFunctions.R
\name{pb_index}
\alias{pb_index}
\title{Presence and Balance Indices -- pb_index()}
\usage{
pb_index(wimp, method = "wnorm", std = "none")
}
\arguments{
\item{wimp}{A wimp object containing implication grid and constructs.}

\item{method}{Method for calculating degree indices. Default is "wnorm".
Options: "wnorm", "simple", "weight", etc.}

\item{std}{Standardization method for P and B indices. Options:
\itemize{
  \item 'none': No standardization (default)
  \item 'vertices': Standardize by max total degree
  \item 'edges': Standardize by total number of edges
  \item 'max_edges': Standardize by max outgoing edges
  \item 'density': Adjust by grid density
}}
}
\value{
A matrix with two columns, 'p' for presence and 'b' for balance,
        containing the indices for each construct. If standardization is
        applied, these values are modified according to the selected method.
}
\description{
Computes presence (P) and balance (B) indices for constructs.
             P represents frequency of occurrence, B represents influence
             balance.
}
\examples{

pb_index(example_wimp)
pb_index(example_wimp, std = "vertices")
pb_index(example_wimp, method = "wnorm", std = "none")

}
\references{
Sanfeliciano, A., Saúl, L. A., Hurtado-Martínez, C., & Botella, L. (2025). PB Space: A Mathematical Framework for Modeling Presence and Implication Balance in Psychological Change Through Fuzzy Cognitive Maps. Axioms.
}
\author{
Carlos Hurtado and Alejandro Sanfeliciano
}
