% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentralityFunctions.R
\name{pb_plot}
\alias{pb_plot}
\title{PB Space Scatter Plot -- pb_plot()}
\usage{
pb_plot(wimp, text_size = 1, lang = "en", ...)
}
\arguments{
\item{wimp}{Subject's WimpGrid. It must be a "wimp" S3 object
imported by the \\code{\\link{importwimp}} function.}

\item{text_size}{Text label size. Default is 1.}

\item{...}{Additional arguments passed to \\code{\\link{pb_index}} function.}
}
\value{
A Plotly object representing the generated scatter plot.
}
\description{
Creates a scatter plot of constructs in Presence-Balance
             space. P represents construct frequency, B represents
             construct influence balance.
}
\examples{

pb_plot(example_wimp)

}
\references{
Sanfeliciano, A., Saúl, L. A., Hurtado-Martínez, C., & Botella, L. (2025). PB Space: A Mathematical Framework for Modeling Presence and Implication Balance in Psychological Change Through Fuzzy Cognitive Maps. Axioms.
}
\author{
Carlos Hurtado and Alejandro Sanfeliciano
}
