% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemDynamicsFunctions.R
\name{pcsd_summary}
\alias{pcsd_summary}
\title{PCSD summary -- pcsd_summary()}
\usage{
pcsd_summary(scn)
}
\arguments{
\item{scn}{A "scn" object returned by \code{scenariomatrix()}.}
}
\value{
A matrix with PCSD summary metrics per construct.
}
\description{
Summarize PCSD results: initial, final, average, and change per
  construct.
}
\examples{
# Example Scenario Matrix (scn)
example_scn <- scenariomatrix(
  wimp = example_wimp,
  infer = "self dynamics",
  thr = "saturation",
  act_vector = c(0.1, -0.2, 0.3, -0.1, 0)
)

# Calculate PCSD summary for each construct
pcsd_summary(example_scn)

}
