% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentralityFunctions.R
\name{ph_plot}
\alias{ph_plot}
\title{Scatter Plot of Constructs in PH Space -- ph_plot()}
\usage{
ph_plot(wimp, text.size = 1, ...)
}
\arguments{
\item{wimp}{An object of class 'wimp', which contains an implication grid
and associated constructs.}

\item{text.size}{Size of the text labels. Default is 1.}

\item{...}{Additional arguments are passed from \code{\link{ph_index}} function.}
}
\value{
A Plotly object representing the generated scatter plot.
}
\description{
This function generates a scatter plot of constructs in the P-H space,
where P represents Presence (frequency of the construct) and H represents Hierarchy
(influence of the construct).
}
\examples{

ph_plot(example.wimp)

}
\author{
Carlos Hurtado and Alejandro Sanfeliciano
}
