% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Methods.R
\name{print.scn}
\alias{print.scn}
\title{Print method for scn objects}
\usage{
\method{print}{scn}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class `scn` returned by `scenariomatrix()`.}

\item{digits}{Digits to round numeric values. Default: 3.}

\item{...}{Ignored.}
}
\description{
Shows the parameters used to generate the scenario matrix and the values
for each iteration.
}
