% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Methods.R
\name{print.wimp}
\alias{print.wimp}
\title{Print method for wimp objects}
\usage{
\method{print}{wimp}(x, n = 10, m = 10, digits = 3, ...)
}
\arguments{
\item{x}{A `wimp` object created by `importwimp()`.}

\item{n}{Number of vertex rows to show (head). Default: 10.}

\item{m}{Number of edge rows to show (head). Default: 10.}

\item{digits}{Number of digits for numeric rounding. Default: 3.}

\item{...}{Ignored.}
}
\description{
Presents a concise, readable summary of a `wimp` object, including
global metadata, a preview of vertices, and a preview of edges.
}
\examples{
# print(example_wimp)

}
