% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtCFunctions.R
\name{rtc2_plot}
\alias{rtc2_plot}
\title{Visualization of the negative effects of Consequences and Feedback}
\usage{
rtc2_plot(wimp, std = FALSE)
}
\arguments{
\item{wimp}{An object of class 'wimp', which contains an implication grid and associated constructs.}

\item{std}{A logical value indicating whether to standarize the scores (TRUE) or (FALSE)}
}
\value{
returns a scatter plot showing the negative effects of consequences and feedback on each construct of the system
}
\description{
This function creates a scatter plot to show the negative effects of the consequences and feedbacks of each construct to analyze their resistance to change
}
\examples{
rtc2_plot (example.wimp)
}
