% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustmentFunctions.R
\name{self_plot}
\alias{self_plot}
\title{Self-Ideal Radar Chart -- self_plot()}
\usage{
self_plot(wimp)
}
\arguments{
\item{wimp}{Subject's WimpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}
}
\value{
Interactive Plotly radar chart with self (blue) and ideal (green)
        traces, including SSI Index in the legend.
}
\description{
Creates a radar chart displaying self and ideal ratings
             for each construct, showing congruence patterns visually.
}
\examples{
self_plot(example_wimp)

}
\author{
Alejandro Sanfeliciano
}
