% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemDynamicsFunctions.R
\name{stability_index}
\alias{stability_index}
\title{PCSD Stability Index -- stability_index()}
\usage{
stability_index(scn)
}
\arguments{
\item{scn}{A "scn" object returned by \code{scenariomatrix()}.}
}
\value{
A vector with the standard deviation of each construct.
}
\description{
Compute the standard deviation of each construct across PCSD
  iterations.
}
\examples{
# Example Scenario Matrix (scn)
example_scn <- scenariomatrix(
  wimp = example_wimp,
  infer = "self dynamics",
  thr = "saturation",
  act_vector = c(0.1, -0.2, 0.3, -0.1, 0)
)

# Calculate stability index for each construct
stability_index(example_scn)

}
