% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R
\docType{data}
\name{su.wimp}
\alias{su.wimp}
\title{Example Subject WimpGrid}
\format{
An object of class \code{wimp} (inherits from \code{list}) of length 6.
}
\usage{
su.wimp
}
\description{
A sample "wimp" object representing data from SU clinical case. This dataset
             is used for illustrating functions and testing purposes.
}
\examples{
digraph(su.wimp)
}
\keyword{datasets}
