% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R
\docType{data}
\name{su_wimp}
\alias{su_wimp}
\title{SU Clinical Case WimpGrid}
\format{
An object of class \code{wimp} (inherits from \code{list}) of length 3.
}
\usage{
su_wimp
}
\description{
A sample "wimp" object representing data from SU clinical case.
             This dataset is used for illustrating functions and testing
             purposes.
}
\examples{
digraph(su_wimp)
}
\keyword{datasets}
