% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentralityFunctions.R
\name{test_optimal_num_clusters}
\alias{test_optimal_num_clusters}
\title{Public test function for .optimal.num.clusters hide function}
\usage{
test_optimal_num_clusters(...)
}
\description{
This function computes the presence (P, frequency of occurrence) and
}
