% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widgets.R
\name{widget_adjustment}
\alias{widget_adjustment}
\title{Generate Tabbed HTML Widget for Wellness Analysis (Análisis de Bienestar) / Adjustment}
\usage{
widget_adjustment(x, y = NULL, lang = "en", ...)
}
\arguments{
\item{x}{A `wimp` object representing the baseline evaluation.}

\item{y}{An optional `wimp` object representing the post evaluation.}

\item{lang}{Language parameter ("en" or "es").}

\item{...}{Additional arguments passed to the underlying plot functions.}
}
\value{
An object of class `htmltools::browsable`.
}
\description{
Creates a standalone HTML widget containing a two-tab interface for analyzing wellness.
If one WimpGrid is provided, it shows the Self analysis and SSI structure.
If two WimpGrids are provided, it shows the Self monitoring and SSI monitoring.
}
\examples{
\dontrun{
widget_adjustment(example_wimp)
widget_adjustment(example_wimp, example_wimp_post)
}
}
