% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemDynamicsFunctions.R
\name{auc_index}
\alias{auc_index}
\title{PCSD AUC Index -- auc_index()}
\usage{
auc_index(scn, std = TRUE)
}
\arguments{
\item{scn}{A "scn" S3 class object, the result of the \code{\link{scenariomatrix}} function.
This object contains the scenario matrix and relevant metadata.}

\item{std}{A logical value. If \code{TRUE}, standardizes the AUC results.}
}
\value{
Returns a vector with the AUC index of each construct.
}
\description{
This function calculates the area under the PCSD curve for each
construct.
}
\examples{
# Example Scenario Matrix (scn)
example_scn <- scenariomatrix(
  wimp = example.wimp,
  infer = "self dynamics",
  thr = "saturation",
  act.vector = c(0.1, -0.2, 0.3, -0.1, 0)
)

# Calculate AUC index for each construct
auc_index(example_scn,std = TRUE)

}
