% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtCFunctions.R
\name{cq.fb_index}
\alias{cq.fb_index}
\title{Calculation of consequence and feedback effects for each construct of the
system}
\usage{
cq.fb_index(wimp, std = FALSE)
}
\arguments{
\item{wimp}{An object of class 'wimp', which contains an implication grid and associated constructs.}

\item{std}{A logical value indicating whether to standarize the scores (TRUE) or (FALSE)}
}
\value{
Returns a dataframe with the intensity of the effects caused by consequences and feedback interactions
}
\description{
This function calculates the intensity of the effects of consequences and feedbacks
for each construct to analyze its resistance to change
}
\examples{
cq.fb_index (example.wimp)

}
