% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtCFunctions.R
\name{cq.fb_index}
\alias{cq.fb_index}
\title{Calculation of consequence and feedback effects for each construct of the
system}
\usage{
cq.fb_index(wimp, std = FALSE)
}
\arguments{
\item{std}{}
}
\value{
Returns a dataframe with numbers of consequences and feedbacks
}
\description{
This function calculates the numbers of consequences and feedbacks
for each construct to analyze its resistance to change
}
\examples{
cq.fb_index (example.wimp)

}
