% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtCFunctions.R
\name{cq/fb_index}
\alias{cq/fb_index}
\title{Calculation consequence and feedback}
\usage{
cq/fb_index(wimp, std = FALSE)
}
\arguments{
\item{std}{}
}
\description{
Calculation consequence and feedback
}
\examples{
cq/fb_index (example.wimp)

}
