% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{digraph}
\alias{digraph}
\title{Selfdigraph -- digraph()}
\usage{
digraph(
  wimp,
  vertex.vector = NA,
  ideal.vector = NA,
  width = "100\%",
  height = "1000px",
  color = "red/green",
  layout = "graphopt",
  show = TRUE,
  hide.inverse = FALSE
)
}
\arguments{
\item{hide.inverse}{}
}
\description{
Selfdigraph -- digraph()
}
