% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentralityFunctions.R
\name{graph_ph}
\alias{graph_ph}
\title{Scatter Plot of Constructs in P-H Space}
\usage{
graph_ph(..., mark.nva = TRUE, mark.cnt = TRUE, show.points = TRUE)
}
\arguments{
\item{mark.nva}{Boolean value that specifies if non-viable areas are to be marked in the
graphic. Non-viable areas are parts of the P-H space where constructs cannot logically exist.}

\item{mark.cnt}{Boolean value that specifies if central constructs have to be highlighted.
Central constructs are depicted with a distinct color and symbol to differentiate them from
peripheral constructs.}

\item{show.points}{Boolean value that specifies whether points should be displayed or not}

\item{phm.mat}{A matrix where each row represents a construct and contains
the P and H coordinates of the construct, as well as an indication of whether the
construct is central (1) or not (0). The matrix must have row names,
which are used to label the constructs in the graph.}
}
\value{
A Plotly object representing the generated scatter plot.
}
\description{
This function generates a scatter plot of constructs in the P-H space,
where P represents Presence (frequency of the construct) and H represents Hierarchy
(influence of the construct). Central constructs are highlighted in red color, and
peripheral constructs in another, facilitating their visual identification.
}
\examples{
graph_ph(phm.mat, mark.nva = TRUE, mark.cnt = TRUE)
}
