% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustmentFunctions.R
\name{hypo_plot}
\alias{hypo_plot}
\title{Hypothetical Scenarios Plot  -- hypo_plot()}
\usage{
hypo_plot(wimp, text.size = 1, ...)
}
\arguments{
\item{wimp}{Subject's WimpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}

\item{text.size}{Scalar that modifies the text size. Default is 1.}

\item{...}{additional arguments are passed from \code{\link{self_index}}
function.}

\item{center}{Establishes the centre of the frame. Use "data" to set the data
to be framed and "origin" to set the origin to be in the centre. the default
is "data".}
}
\value{
returns a interactive scatter plot made with Plotly.
}
\description{
This function creates a scatter plot to show the results of the
             \code{\link{self_index}} function.
}
\examples{

hypo_plot (example.wimp)
}
\author{
Maite Benitez Santos, Guillermo Calleja Garate and Alejandro Sanfeliciano
}
