% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{idealdigraph}
\alias{idealdigraph}
\title{Ideal digraph -- idealdigraph()}
\usage{
idealdigraph(wimp, hide_direct = FALSE, layout = "circle", ...)
}
\arguments{
\item{wimp}{Subject's WimpGrid. It must be a "wimp" S3 object
imported by the \\code{\\link{importwimp}} function.}

\item{hide_direct}{If TRUE, hide direct relationships between nodes. Default is
FALSE.}

\item{layout}{Layout for the digraph. Options: "circle", "rtcircle", "tree",
"graphopt", "mds", "grid" or "areas". Default is "circle".}

\item{...}{Additional arguments passed to \\code{\\link{digraph}}.}
}
\value{
A visNetwork graph
}
\description{
Plot the ideal self based on the constructs and their relations.
}
\examples{
idealdigraph(example_wimp)
}
\author{
Alejandro Sanfeliciano
}
