% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{idealdigraph}
\alias{idealdigraph}
\title{Ideal digraph -- idealdigraph()}
\usage{
idealdigraph(wimp, inc = TRUE, layout = "circle", ...)
}
\arguments{
\item{layout}{}
}
\description{
Ideal digraph -- idealdigraph()
}
