% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeImplicationsFunctions.R
\name{if_barchart}
\alias{if_barchart}
\title{Impact and Feedback Bar Chart - if_barchart()}
\usage{
if_barchart(wimp, show = "all", ...)
}
\arguments{
\item{wimp}{Subject's WimpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}

\item{show}{Construct filter. One of `"all"`, `"dil"`, `"nodil"`.}

\item{...}{Additional arguments forwarded to \code{if_index()}.}
}
\value{
A plotly subplot with Impact and Feedback panels.
}
\description{
Two-panel horizontal bar chart showing positive and
negative components of Impact and Feedback.
}
\details{
Dilemmatic constructs (ideal = 0) are highlighted in yellow
  when included.
}
\examples{
if_barchart(example_wimp)
}
\author{
Alejandro Sanfeliciano
}
