% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeImplicationsFunctions.R
\name{if_barchart}
\alias{if_barchart}
\title{Impact and Feedback Bar Chart  -- if_barchart()}
\usage{
if_barchart(wimp, show = "all", ...)
}
\arguments{
\item{wimp}{Subject's WimpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}

\item{show}{A character string indicating the constructs to be displayed. 'all' will display
all constructus, 'dil' will only display dilemmatic constructs and 'nodil' will
exclude dilemmatic constructs. Default is 'all'.}

\item{...}{additional arguments are passed from \code{\link{if_index}}
function.}
}
\value{
returns a interactive bar chart made with plotly.
}
\description{
This function creates a bar chart to show the results of the
             \code{\link{if_index}} function.
}
\examples{

if_barchart(example.wimp)

}
\author{
Alejandro Sanfeliciano
}
