% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeImplicationsFunctions.R
\name{if_index}
\alias{if_index}
\title{Impact And Feedback Index -- if_index()}
\usage{
if_index(wimp, std = "adjacent")
}
\arguments{
\item{wimp}{An object of class 'wimp', which contains a Weigthed Implication
Grid.}

\item{std}{A character string indicating the type of standardization to apply.
Options are 'none' (no standardization), 'vertex' (standardize by the maximum total degree
of the constructs), 'edges' (standardize by the total number of edges) and 'adjacent'
(standarized by the number of the adjacent vertices). Default is 'adjacent'.}
}
\value{
Returns a dataframe showing the intensity of the impact in the
system and feedback from the system for each construct (negative, postive and
 global).
}
\description{
This function calculates the intensity of the of impact in the
system and feedback from the system for each construct.
}
\examples{

if_index (example.wimp)

}
\author{
Maite Benitez Santos, Guillermo Calleja Garate, Alejandro Sanfeliciano
}
