% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonitoringFunctions.R
\name{monitoring_adj}
\alias{monitoring_adj}
\title{Monitoring Self Adjustment -- monitoring_adj()}
\usage{
monitoring_adj(wimp.t0, wimp.t1, legend = TRUE)
}
\arguments{
\item{wimp.t0}{First Subject's WimpGrid (e.g., pre-intervention). It must be a "wimp" S3 object.}

\item{wimp.t1}{Second Subject's WimpGrid (e.g., post-intervention). It must be a "wimp" S3 object.}

\item{legend}{If TRUE, displays legend of plot. Default is TRUE.}
}
\value{
Interactive radar plot using plotly
}
\description{
This function generates an interactive radar plot to visualize
and compare the progress of the "self" and "ideal" across different
time points (e.g., pre- and post-intervention).
}
\examples{

 monitoring_adj(example.wimp,example.wimp)
}
