% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonitoringFunctions.R
\name{monitoring_ph}
\alias{monitoring_ph}
\title{Monitoring PH index -- monitoring_ph()}
\usage{
monitoring_ph(wimp.t0, wimp.t1, show.centroid = TRUE, text.size = 1, ...)
}
\arguments{
\item{wimp.t0}{Data object containing constructs and their respective P and H coordinates for the first grid.}

\item{wimp.t1}{Data object containing constructs and their respective P and H coordinates for the second grid}

\item{show.centroid}{Logical; if TRUE, displays the centroid of construct P-H coordinates for both grids on the graph.}

\item{text.size}{Size of the text labels. Default is 1.}

\item{...}{additional arguments are passed from \code{\link{ph_index}} function.}
}
\value{
A Plotly object representing the comparative graph of constructs across both grids
}
\description{
This function generates a graphical comparison of constructs from two
different Weigthed Implications Grids plotted in a PH space.
}
\examples{

monitoring_ph(example.wimp,example.wimp)
}
