% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonitoringFunctions.R
\name{monitoring_ssi}
\alias{monitoring_ssi}
\title{Monitoring SSI Adjustment -- monitoring_ssi()}
\usage{
monitoring_ssi(wimp_t0, wimp_t1)
}
\arguments{
\item{wimp_t0}{First Subject's WimpGrid (e.g., pre-intervention). It must
be a "wimp" S3 object.}

\item{wimp_t1}{Second Subject's WimpGrid (e.g., post-intervention). It must
be a "wimp" S3 object.}
}
\value{
A two heatmap made with plotly.
}
\description{
This function generates an interactive heatmap to visualize
and compare the progress of the SSI index across different
time points (e.g., pre- and post-intervention).
}
\examples{
 monitoring_ssi(example_wimp, example_wimp)
}
