% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentralityFunctions.R
\name{pb_plot}
\alias{pb_plot}
\title{PB Space Scatter Plot -- ph_plot()}
\usage{
pb_plot(wimp, text_size = 1, ...)
}
\arguments{
\item{wimp}{Subject's WimpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}

\item{text_size}{Text label size. Default is 1.}

\item{...}{Additional arguments passed to \code{\link{ph_index}} function.}
}
\value{
A Plotly object representing the generated scatter plot.
}
\description{
Creates a scatter plot of constructs in Presence-Balance
             space. P represents construct frequency, B represents
             construct influence balance.
}
\examples{

pb_plot(example_wimp)

}
\author{
Carlos Hurtado and Alejandro Sanfeliciano
}
