% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemDynamicsFunctions.R
\name{pcsd}
\alias{pcsd}
\title{Personal Constructs System Dynamics plot -- pcsd()}
\usage{
pcsd(scn, vline = NA)
}
\arguments{
\item{scn}{A "scn" S3 class object, the result of the \code{\link{scenariomatrix}} function.
This object contains the scenario matrix and relevant metadata.}

\item{vline}{An optional numeric value indicating a specific iteration to highlight
with a vertical line on the plot. Defaults is NA.}
}
\value{
Interactive plot created with plotly.
}
\description{
Interactive line plot of personal constructs system dinamics.
Show \code{\link{scenariomatrix}} values expressed in terms of distance to
Ideal-Self for each personal construct across the mathematical iterations.
}
\examples{

# Example Scenario Matrix (scn)
example_scn <- scenariomatrix(
  wimp = example.wimp,
  infer = "self dynamics",
  thr = "saturation",
  act.vector = c(0.1, -0.2, 0.3, -0.1, 0)
)

# Plot the dynamics
pcsd(example_scn)

}
