% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtCFunctions.R
\name{rtc_plot}
\alias{rtc_plot}
\title{Visualization of the effects of Consequences and Feedback}
\usage{
rtc_plot(wimp, std = FALSE)
}
\arguments{
\item{wimp}{An object of class 'wimp', which contains an implication grid and associated constructs.}

\item{std}{A logical value indicating whether to standarize the scores (TRUE) or (FALSE)}
}
\value{
returns a scatter plot showing the effects of consequences and feedback on each construct of the system
}
\description{
This function creates a scatter plot to show the effects of the consequences and feedbacks of each construct to analyze their resistance to change
}
\examples{
rtc_plot (example.wimp)
}
