% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemDynamicsFunctions.R
\name{scenariomatrix}
\alias{scenariomatrix}
\title{Scenario Matrix -- scenariomatrix()}
\usage{
scenariomatrix(
  wimp,
  act.vector = NA,
  infer = "self dynamics",
  thr = "saturation",
  max.iter = 5,
  e = 1e-04,
  stop.iter = 2,
  exclude.dilemmatics = FALSE
)
}
\arguments{
\item{wimp}{A "wimp" S3 object.}

\item{act.vector}{A numeric vector representing the actions or interventions
to be applied to the constructs in the WimpGrid.}

\item{infer}{A character string indicating the type of inference to be used
in the scenario analysis. Possible values include "self dynamics"
or "impact dynamics". Default is "self dynamics"}

\item{thr}{A character string specifying the threshold method to be applied
to the constructs. Supported options include "saturation","tanh"
or "none". Default is "saturation"}

\item{max.iter}{An integer specifying the maximum number of iterations for
the scenario analysis. Defaults is 5.}

\item{e}{A numeric value representing the convergence threshold. Smaller
values lead to stricter convergence criteria. Default is 0.0001}

\item{stop.iter}{An integer specifying the number of consecutive iterations
with no significant change required to declare convergence.
Default is 2}

\item{exclude.dilemmatics}{A logical value. If TRUE, dilemmatic constructs
are excluded from the scenario matrix calculations.
Default is TRUE}
}
\value{
A "scn" S3 class object.
}
\description{
This function generates a scenario matrix based on the provided
             WimpGrid object and a vector of actions. It models potential
             outcomes iteratively based on user-defined parameters, allowing
             for scenario analysis and decision-making.
}
\examples{

# Activation vector (example)
act.vector <- c(0.1, -0.2, 0.3, -0.1, 0)

# Generate a scenario matrix
scenariomatrix(wimp = example.wimp, act.vector = act.vector)

}
