% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustmentFunctions.R
\name{self_index}
\alias{self_index}
\title{Correlations between Self and Hypothetical scenarios -- self_index()}
\usage{
self_index(wimp, method = "sim", rc = TRUE, alpha = 0.5, beta = 0.5)
}
\arguments{
\item{wimp}{Subject's WimpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}

\item{method}{The correlation method to use. All methods of the \code{\link{cor}} function
are allowed and "sim" for Tversky similarity. Default is "sim".}

\item{rc}{Use Cohen's rc which is invariant to construct reflection. Default is TRUE.}

\item{alpha}{Alpha value for Tversky similarity calculation. Default is .5.}

\item{beta}{Beta value for Tversky similarity calculation. Default is .5.}
}
\value{
A list with global adjustment self indices and specific indices for each construct.
}
\description{
this function Calculates the global and specific adjustment
             indices of the self for each hypothetical scenario in the WimpGrid.
}
\examples{

self_index(example.wimp)

}
\author{
Alejandro Sanfeliciano
}
