% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{simdigraph}
\alias{simdigraph}
\title{Simulation digraph -- simdigraph()}
\usage{
simdigraph(scn, ...)
}
\arguments{
\item{scn}{A scenario matrix or wimp object.}

\item{...}{Additional arguments}
}
\description{
Plot the hypothetical self for a given scenario iteration.
}
