% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{simdigraph}
\alias{simdigraph}
\title{Simulation digraph -- simdigraph()}
\usage{
simdigraph(scn, niter = 0, ...)
}
\arguments{
\item{scn}{A scenario matrix. It must be a "scn" S3 object
from the \code{\link{scenariomatrix}} function.}

\item{niter}{Iteration displayed in the digraph. Default is 0.}

\item{...}{Additional arguments are passed from \code{\link{digraph}}
function.}
}
\value{
A digraph made with visNetwork
}
\description{
A digraph that represents the hypothetical self of the person being assessed
             on the basis of its constructs and the relationships between them.
}
\examples{

 scn <- scenariomatrix(example.wimp, rep(1,5))
 simdigraph(scn, niter = 2)

}
\author{
Alejandro Sanfeliciano
}
