% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustmentFunctions.R
\name{ssi_heatmap}
\alias{ssi_heatmap}
\title{SSI Heatmap -- ssi_heatmap()}
\usage{
ssi_heatmap(wimp)
}
\arguments{
\item{wimp}{Subject's WimpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}
}
\value{
A plotly heatmap.
}
\description{
Creates a heatmap showing SSI values across different
             alpha (discrepancy salience) and beta (aspiration salience)
             parameter combinations.
}
\examples{
ssi_heatmap(example_wimp)

}
\author{
Alejandro Sanfeliciano
}
