% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemDynamicsFunctions.R
\name{stability_index}
\alias{stability_index}
\title{PCSD Stability Index -- stability_index()}
\usage{
stability_index(scn)
}
\arguments{
\item{scn}{A "scn" S3 class object, the result of the \code{\link{scenariomatrix}} function.
This object contains the scenario matrix and relevant metadata.}
}
\value{
Returns a vector with the standard deviation of each of the
constructs.
}
\description{
This function returns the standard deviation for each
construct over the mathematical iterations of the PCSD.
}
\examples{
# Example Scenario Matrix (scn)
example_scn <- scenariomatrix(
  wimp = example.wimp,
  infer = "self dynamics",
  thr = "saturation",
  act.vector = c(0.1, -0.2, 0.3, -0.1, 0)
)

# Calculate stability index for each construct
stability_index(example_scn)

}
