% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{weight_heatmap}
\alias{weight_heatmap}
\title{Weight Matrix Heatmap -- weight_heatmap()}
\usage{
weight_heatmap(wimp, palette = "Redgreen")
}
\arguments{
\item{wimp}{Subject's WimpGrid object. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}

\item{palette}{Character. Color palette for the heatmap. Options: \code{"Redgreen"}
(default, red-white-green), \code{"viridis"}, \code{"plasma"}, etc.}
}
\value{
A plotly heatmap object.
}
\description{
Creates an interactive plotly heatmap of the weight matrix, 
             reoriented towards the ideal self poles.
}
\examples{
weight_heatmap(example_wimp)

}
\author{
Alejandro Sanfeliciano
}
