% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAFunctions.R
\name{wimp_biplot}
\alias{wimp_biplot}
\title{Weigthed Implication Grid Biplot}
\usage{
wimp_biplot(wimp, text.size = 1)
}
\arguments{
\item{wimp}{Subject's WimpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}

\item{text.size}{Size of the text labels. Default is 1.}
}
\value{
A interactive biplot made with plotly
}
\description{
This function generates a biplot visualization from the
             semi-structured interview data collected via the WimpGrid
             methodology. A biplot is a graphical representation that
             combines both hipothetical selves and personal constructs of a
             matrix into a single plot, making it easier to identify
             patterns, relationships, and clusters within the data.
}
\examples{

wimp_biplot(example.wimp)

}
